feat: Included INHERITANCE logic
Customer and TieredCustomer. Also I removed duplicate data in Customers.csv

Please note that i will be sending the newest customer file to the project info group for you all to import

+++ Backend/timperio.crm/src/main/java/oopsies/timperio/crm/Controller/PurchaseHistoryController.java
@@ -28,7 +28,7 @@ public class PurchaseHistoryController {
 
     // New endpoint to get purchase history by customerId
     @GetMapping("/byCustomer")
-    public ResponseEntity<List<PurchaseHistory>> getPurchaseHistoryByCustomerId(@RequestParam Long customerId) {
+    public ResponseEntity<List<PurchaseHistory>> getPurchaseHistoryByCustomerId(@RequestParam Integer customerId) {
         List<PurchaseHistory> purchaseHistories = purchaseHistoryService.findByCustomerId(customerId);
         return new ResponseEntity<>(purchaseHistories, HttpStatus.OK);
     }
